requirements.txt
Flask[async]==2.0.2
Flask-JWT-Extended==4.3.1
Flask-SQLAlchemy==2.5.1
uvicorn==0.17.1
gunicorn==20.1.0
flasgger==0.9.5
psycopg2-binary==2.9.2
pydantic==1.9.0
pytest==7.0.0
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.27.0